[Add & Fix] Add molecules HeaderUserMenu and fix relation atoms
